Have bandit logger separate out context (#5)
* work in progress

* math in place

* working on tests still

* tests passing

* new test cases

* wip manually verifying math

* tests passing

* remove logging

* clean up score computation

* changes from self-reivew of PR

* consolidate test bandit rac setup code

* clean up test file

* update comments

* more comment improvements

* failing tests in place

* break out attributes by type

* changes from self-review of PR

* return type as provided

* make EppoValue constructor private

* feedback from PR

+++ src/main/java/com/eppo/sdk/helpers/VariationHelper.java
@@ -22,7 +22,7 @@ public class VariationHelper {
         return variation.get();
     }
 
-    static public float variationProbability(Variation variation, int subjectShards) {
-        return (float)(variation.getShardRange().end - variation.getShardRange().start + 1) / subjectShards;
+    static public double variationProbability(Variation variation, int subjectShards) {
+        return (double)(variation.getShardRange().end - variation.getShardRange().start + 1) / subjectShards;
     }
 }
